Vehicle Counting and Detection using OpenCV
Project is about vehicle counting, speed detection, and detect the direction. it is an industry academia collaborated project with PI-LABS Bangladesh Ltd. 
The system can detect the speed of a moving object like vehicle or human, can count the number of moving object and also able to detect the direction.
i have use Tkinter(Python Gui) for developing the interface of the system and use OpenCV to develop back end.

i try to do documentation in my project report.